refactor: create a utility class
